backend/requirements.txt
# Django Core
Django==5.2.6
djangorestframework==3.16.1
django-cors-headers==4.9.0
django-filter==25.1
Pillow==11.3.0

# Database
psycopg2-binary==2.9.10
dj-database-url==2.2.0

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-allauth==64.2.1

# Caching & Sessions
redis==5.2.0
django-redis==5.4.0

# Background Tasks
django-celery-beat==2.8.0
flower==2.0.1

# Distributed Coordination
kazoo>=2.9.0

# Event Streaming & Message Queue
kafka-python>=2.0.2
confluent-kafka>=2.3.0
aiokafka>=0.8.10

# Message Broker & Queue
pika>=1.3.2
celery[redis,rabbitmq]>=5.4.0
kombu>=5.3.4
amqp>=5.2.0

# Email & Templates
django-anymail>=10.2
premailer>=3.10.0
beautifulsoup4>=4.12.2
lxml>=4.9.3
html2text>=2020.1.16

# File Storage (Local/Alternative Cloud Storage)
django-storages==1.14.4
# Removed: boto3==1.35.40  # AWS SDK - removed for non-AWS deployment

# API Documentation
drf-spectacular==0.27.2

# Production Server
gunicorn==23.0.0
gevent==24.2.1

# Monitoring & Logging
sentry-sdk==2.17.0
django-extensions==3.2.3

# Environment Management
python-decouple==3.8
django-environ==0.11.2

# Testing
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==5.0.0

# Utilities
requests==2.32.3
python-dateutil==2.9.0

# OpenTelemetry and Observability
-r requirements-opentelemetry.txt

# Apache Spark
pyspark>=3.5.0
findspark>=2.0.1
 
# Monitoring
prometheus-client>=0.17.0
influxdb>=5.3.1